Refactor ProcessManager events.
